feat: add NotFound page and refactor pages folder
