--- src/pages/Home.module.css
@@ -1,15 +1,0 @@
-.main {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 20px;
-}
-
-.mainTitle {
-    font-size: 60px;
-    background: linear-gradient(#ffffff 50%, #000000 100%);
-    background-clip: text;
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-}
