+++ src/pages/Home/Home.jsx
@@ -1,11 +1,11 @@
 import style from "./Home.module.css";
-import { Logo } from "../components";
-import Button from "../components/Button/Button";
+import { Logo } from "../../components";
+import Button from "../../components/Button/Button";
 
 export default function Home() {
   return (
     <>
-      <main className={style.main}>
+      <main>
         <Logo size={150} />
         <h1 className={style.mainTitle}>TicTacToe</h1>
         <Button to="/play">Jogar</Button>
