+++ src/pages/Home/Home.module.css
@@ -1,0 +1,7 @@
+.mainTitle {
+  font-size: 60px;
+  background: linear-gradient(#ffffff 50%, #000000 100%);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
