+++ src/pages/index.js
@@ -1,2 +1,3 @@
-export { default as Home } from "./Home"
-export { default as Play } from "./Play"
+export { default as Home } from "./Home/Home";
+export { default as Play } from "./Play/Play";
+export { default as NotFound } from "./NotFound/NotFound";
